moving script to query.py, adding exception, and adding .ndjson extension to temp file
